made selected movie stays selected after rating
updated "handleClick" so the movie selected on the tableview will stay selected after rating. This prevents issues where clicking on the same movie while it's still playing after rating will make the same movie play twice. That may be a larger issue that may need to be fixed but for now, this is a solid way of fixing the issue